% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_export.R
\name{report_export}
\alias{report_export}
\title{Export list of data frames to separate excel tabs}
\usage{
report_export(
  tabs,
  export_dest = "report_exports",
  fname = "export",
  export_time = TRUE
)
}
\arguments{
\item{tabs}{data frames to export}

\item{export_dest}{export destination}

\item{fname}{export file name}

\item{export_time}{logical. If true the current time is added to the file name}
}
\value{
none
}
\description{
Supply list of data frames and export to an excel file
in an existing or new directory with optional timestamp
}
\examples{
\dontrun{
report_export(list(mtcars, airquality), export_dest = "new_reports", fname = "curr_report")
}
}
